--- packages/server/customer-os-common-module/service/email_in_analyzer_types.go
@@ -1,8 +1,0 @@
-package service
-
-type HeaderAnalysis struct {
-	ProcessEmail    bool
-	IsBounce        bool
-	IsAutoResponder bool
-	IsBulkMail      bool
-}
